+++ core/src/com/tint/specular/upgrades/PDSUpgrade.java
@@ -11,10 +11,10 @@ public class PDSUpgrade extends Upgrade {
 	private AtlasRegion tex;
 	
 	public PDSUpgrade(GameState gs) {
-		super(gs, Specular.prefs.getInteger("PDS Upgrade Grade"), 10);
+		super(gs, Specular.prefs.getFloat("PDS Upgrade Grade"), 10);
 	}
 	
-	public PDSUpgrade(int grade, int maxGrade, TextureAtlas ta) {
+	public PDSUpgrade(float grade, float maxGrade, TextureAtlas ta) {
 		super(null, grade, maxGrade);
 		tex = ta.findRegion("game1/Point Defense");
 	}
